+++ example/dart_ping_example.dart
@@ -12,15 +12,4 @@ void main() async {
   ping.stream.listen((event) {
     print(event);
   });
-
-  // [Optional]
-  // Waiting for ping to output first two results
-  // Not needed in actual use. For example only!
-  await Future.delayed(Duration(seconds: 2));
-
-  // [Optional]
-  // Stop the ping prematurely and output a summary
-  // Make sure you do not call this before listening to the stream!
-  // Not necessary for actual use. For example only!
-  await ping.stop();
 }