+++ lib/src/native_ping.dart
@@ -7,7 +7,7 @@ import 'package:dart_ping/src/ping/linux_ping.dart';
 import 'package:dart_ping/src/ping/mac_ping.dart';
 import 'package:dart_ping/src/ping/windows_ping.dart';
 
-Ping getPing(String host, int? count, double interval, double timeout, int ttl,
+Ping getPing(String host, int? count, int interval, int timeout, int ttl,
     bool ipv6, PingParser? parser, Encoding encoding) {
   switch (Platform.operatingSystem) {
     case 'android':